move telemetry down into redux context
